pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calipy"
version = "0.5.0"
description = "CaliPy, the Calibration library Python. A library for building and solving probabilistic instrument models."
readme = "README.md"
requires-python = ">=3.7"
license = { text = "Prosperity Public License" }

authors = [
  { name = "Dr. Jemil Avers Butt", email = "jemil.butt@atlasoptimization.com" }
]

classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Mathematics",
  "License :: Other/Proprietary License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10"
]

dependencies = [
  "torch>=2.0",
  "pyro-ppl>=1.8.6",
  "pandas>=2.0.0",
  "matplotlib>=3.3.0",
  "einops>=0.8.1",
  "torchviz>=0.0.2",
  "varname>=0.8.1",
  "networkx>=2.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=3.7"
]

[project.urls]
Homepage = "http://github.com/atlasoptimization/calipy"
Documentation = "https://your-docs-site.readthedocs.io"
Source = "http://github.com/atlasoptimization/calipy"
Issues = "http://github.com/atlasoptimization/calipy/issues"

